+++ lib/src/algorithms/xxhash64.dart
@@ -1,4 +1,4 @@
 // Copyright (c) 2023, Sudipto Chandra
 // All rights reserved. Check LICENSE file for details.
 
-export 'longs_64bit.dart' if (dart.library.js) 'longs_32bit.dart';
+export 'xxhash64_64bit.dart' if (dart.library.js) 'xxhash64_32bit.dart';