+++ lib/src/hashlib_base.dart
@@ -28,3 +28,4 @@ export 'sha512t.dart';
 export 'shake128.dart';
 export 'shake256.dart';
 export 'xxhash32.dart';
+export 'xxhash64.dart';