+++ test/xxhash64_test.dart
@@ -5,7 +5,7 @@ import 'package:hashlib/hashlib.dart';
 import 'package:test/test.dart';
 
 const seed = 0x9E3779B1;
-const xxh32_1 = XXHash32(seed);
+const xxh64_1 = XXHash64(seed);
 const data = <int>[
   158, 255, 31, 75, 94, 83, 47, 221, 181, 84, 77, 42, 149, 43, 87, 174, 93, //
   186, 116, 233, 211, 166, 76, 152, 48, 96, 192, 128, 0, 0, 0, 0, 0, 0, 0,
@@ -15,34 +15,34 @@ const data = <int>[
 ];
 
 void main() {
-  group('XXHash32 test', () {
+  group('XXHash64 test', () {
     test('with seed = 0 and an empty string', () {
-      expect(xxh32.convert([]).hex(true), "02CC5D05");
-    });
+      expect(xxh64.convert([]).hex(), "ef46db3751d8e999");
+    }, tags: 'skip-js');
     test('with seed = 0 and a single letter', () {
-      expect(xxh32.convert([data[0]]).hex(true), "B85CBEE5");
-    });
+      expect(xxh64.convert([data[0]]).hex(), "4fce394cc88952d8");
+    }, tags: 'skip-js');
     test('with seed = 0 and 14 letters', () {
-      expect(xxh32.convert(data.take(14).toList()).hex(true), "E5AA0AB4");
-    });
+      expect(xxh64.convert(data.take(14).toList()).hex(), "cffa8db881bc3a3d");
+    }, tags: 'skip-js');
     test('with seed = 0 and 101 letters', () {
-      expect(xxh32.convert(data).hex(true), "1F1AA412");
-    });
+      expect(xxh64.convert(data).hex(), "0eab543384f878ad");
+    }, tags: 'skip-js');
 
     test('with seed = $seed and an empty string', () {
-      expect(xxh32_1.convert([]).hex(true), "36B78AE7");
-    });
+      expect(xxh64_1.convert([]).hex(), "ac75fda2929b17ef");
+    }, tags: 'skip-js');
 
     test('with seed = $seed and a single letter', () {
-      expect(xxh32_1.convert([data[0]]).hex(true), "D5845D64");
-    });
+      expect(xxh64_1.convert([data[0]]).hex(), "739840cb819fa723");
+    }, tags: 'skip-js');
 
     test('with seed = $seed and 14 letters', () {
-      expect(xxh32_1.convert(data.take(14).toList()).hex(true), "4481951D");
-    });
+      expect(xxh64_1.convert(data.take(14).toList()).hex(), "5b9611585efcc9cb");
+    }, tags: 'skip-js');
 
     test('with seed = $seed and 101 letters', () {
-      expect(xxh32_1.convert(data).hex(true), "498EC8E2");
-    });
+      expect(xxh64_1.convert(data).hex(), "caa65939306f1e21");
+    }, tags: 'skip-js');
   });
 }